Renamed Frequency.
